+++ src/positioning/CssTranslatePositionProvider.js
@@ -41,19 +41,17 @@ class CssTranslatePositionProvider extends PositionProvider {
    * @override
    * @inheritDoc
    */
-  setPosition(location) {
+  async setPosition(location) {
     ArgumentGuard.notNull(location, "location");
 
-    const that = this;
     this._logger.verbose(`CssTranslatePositionProvider - Setting position to: ${location}`);
 
     const spp = new ScrollPositionProvider(this._logger, this._executor);
-    return spp.setPosition(Location.ZERO).then(() => {
-      return EyesWDIOUtils.translateTo(this._executor, location);
-    }).then(() => {
-      that._logger.verbose("Done!");
-      that._lastSetPosition = location;
-    });
+    await spp.setPosition(Location.ZERO);
+    await this._executor.executeScript(`document.documentElement.style.transform = 'translate(10px, -${location.getY()}px)';`);
+    await EyesWDIOUtils.translateTo(this._executor, location);
+    this._logger.verbose("Done!");
+    this._lastSetPosition = location;
   }
 
   /**